Fix UUID type hints: change userId from int to string
- Updated DuplicateDetectionService to accept string (UUID) for userId
- Updated SemanticSearchController search methods to use string userId
- Removed unnecessary int cast in duplicates page
- UUIDs should remain as strings, not cast to integers

+++ resources/views/livewire/admin/duplicates/index.blade.php
@@ -19,7 +19,7 @@ $search = function () {
     $this->duplicates = [];
 
     $service = app(DuplicateDetectionService::class);
-    $userId = (int) auth()->id();
+    $userId = auth()->id();
 
     $results = match ($this->modelType) {
         'Event' => $service->findDuplicateEvents($userId, $this->threshold, $this->limit),
